added authentication via OpenID
I struggled with testing this. I'm pretty sure I can mock this for testing using VCR and/or fakeweb, but it will take me more time to dig around and find the proper way to do it.

I left the login/logout specs pending. In the meantime, I mocked authentication altogether (i.e. the signed_in? method) for the import request specs.